refactor: 비교하기 페이지를 FSD 구조로 리팩토링
- `app/(main)/compare/page.tsx`: 페이지 엔트리 포인트
- `views/compare-main-page`: 뷰 컴포넌트
- `widgets/compare-list`: 목록 컴포넌트(FavoriteListSection), 하단 바(SelectionBar)
