+++ apps/knockdog/app/(main)/compare/page.tsx
@@ -1,0 +1,5 @@
+import { CompareMainPage } from '@views/compare-main-page';
+
+export default function Page() {
+  return <CompareMainPage />;
+}
