+++ apps/knockdog/src/views/compare-main-page/index.ts
@@ -1,0 +1,1 @@
+export { CompareMainPage } from './ui/CompareMainPage';
