+++ apps/knockdog/src/widgets/compare-list/index.ts
@@ -1,0 +1,2 @@
+export { FavoriteListSection } from './ui/FavoriteListSection';
+export { SelectionBar } from './ui/SelectionBar';
